Format fixes, flow fixes

+++ src/shared/clients/sqs_client.py
@@ -2,6 +2,7 @@ import json
 from mypy_boto3_sqs.client import (
   SQSClient as AwsSqsClient,
 )
+
 from typing import cast, Generic, TypeVar
 
 from shared.clients.aws_client import AwsClient
@@ -23,12 +24,15 @@ class SqsClient(AwsClient, Generic[MessageType]):
       MessageGroupId="callbacks",
     )
 
-  def pop(self) -> MessageType:
+  def pop(self, wait_time: int = 5) -> MessageType:
     response = self.client.receive_message(
-      QueueUrl=self.queue_url, MaxNumberOfMessages=1
+      QueueUrl=self.queue_url, MaxNumberOfMessages=1, WaitTimeSeconds=wait_time
     )
-    message = response["Messages"][0]
-    self.client.delete_message(
-      QueueUrl=self.queue_url, ReceiptHandle=message["ReceiptHandle"]
-    )
-    return cast(MessageType, json.loads(message["Body"]))
+    if len(response["Messages"]):
+      message = response["Messages"][0]
+      self.client.delete_message(
+        QueueUrl=self.queue_url, ReceiptHandle=message["ReceiptHandle"]
+      )
+      return cast(MessageType, json.loads(message["Body"]))
+
+    raise Exception(f"Failed to pop queue item after {wait_time} seconds")
